+++ tests/fixtures/noqa/noqa.py
@@ -712,3 +712,6 @@ for element in range(10):
         # https://github.com/wemake-services/wemake-python-styleguide/issues/1082
         break
     my_print(4)
+
+def raise_bad_exception():
+    raise Exception  # noqa: WPS454