+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Semantic versioning in our case means:
 - Forbids to use single `return None`
 - Add `__await__` to the list of priority magic methods
 - Forbids to use float zeros (`0.0`)
+- Forbids `raise Exception` and `raise BaseException`
 
 ### Bugfixes